+++ NEWS.md
@@ -1,5 +1,8 @@
 # haven 0.1.1.9000
 
+* `read_spss()` uses extension to automatically choose between `read_sav()`
+  and `read_por()` (#53)
+
 * Updates from ReadStat. Including fixes for various parsing bugs, more 
   encodings, and better support for large files.